+++ app/Http/Controllers/InteriorPostController.php
@@ -2,7 +2,7 @@
 
 namespace App\Http\Controllers;
 
-include ("../vendor/autoload.php");
+// include ("../vendor/autoload.php");
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
@@ -12,9 +12,10 @@ use App\ArchitectUploadModel;
 use App\ArchitectProfileModel;
 use App\CheckOutModel;
 use App\User;
-
+use Illuminate\Support\Facades\Hash;
 use ElephantIO\Client;
 use ElephantIO\Engine\SocketIO\Version2x;
+use Image;
 
 class InteriorPostController extends Controller
 {
@@ -77,6 +78,31 @@ class InteriorPostController extends Controller
             $designCode = $random_number .  '-' . $designNumber;
         /*FOR DESIGN CODE END*/
 
+        $thumbnailImage_main = Image::make($file); 
+        $thumbnailImage_thumbnail1 = Image::make($file1); 
+        $thumbnailImage_thumbnail2 = Image::make($file2); 
+        $thumbnailImage_thumbnail3 = Image::make($file3); 
+        $thumbnailImage_thumbnail4 = Image::make($file4); 
+
+        $thumbnailPath_main = storage_path("app/public").'/thumbnail/main_pic/';
+        $thumbnailPath_thumbnail1 = storage_path("app/public").'/thumbnail/thumbnail1/';
+        $thumbnailPath_thumbnail2 = storage_path("app/public").'/thumbnail/thumbnail2/';
+        $thumbnailPath_thumbnail3 = storage_path("app/public").'/thumbnail/thumbnail3/';
+        $thumbnailPath_thumbnail4 = storage_path("app/public").'/thumbnail/thumbnail4/';
+
+        $thumbnailImage_main->resize(150,150);
+        $thumbnailImage_thumbnail1->resize(150,150);
+        $thumbnailImage_thumbnail2->resize(150,150);
+        $thumbnailImage_thumbnail3->resize(150,150);
+        $thumbnailImage_thumbnail4->resize(150,150);
+
+        $thumbnailImage_main->save($thumbnailPath_main.$this->getUserDir().$designCode . '.' . $ext);
+        $thumbnailImage_thumbnail1->save($thumbnailPath_thumbnail1.$this->getUserDir().$designCode . '.' . $ext);
+        $thumbnailImage_thumbnail2->save($thumbnailPath_thumbnail2.$this->getUserDir().$designCode . '.' . $ext);
+        $thumbnailImage_thumbnail3->save($thumbnailPath_thumbnail3.$this->getUserDir().$designCode . '.' . $ext);
+        $thumbnailImage_thumbnail4->save($thumbnailPath_thumbnail4.$this->getUserDir().$designCode . '.' . $ext);
+
+
         $file->storeAs('/public/' . $portfolio . '/' . $main_pic . '/' . $this->getUserDir() . '/' . $type . '/', $designCode . '.' . $ext);    
         $file1->storeAs('/public/' . $portfolio . '/' . $thumbnail1 . '/' . $this->getUserDir() . '/' . $type . '/', $designCode . '.' . $ext);  
         $file2->storeAs('/public/' . $portfolio . '/' . $thumbnail2 . '/' . $this->getUserDir() . '/' . $type . '/', $designCode . '.' . $ext);  
@@ -137,6 +163,7 @@ public function delete_portfolio($id) {
         $file = ArchitectUploadModel::findOrFail($id);
 
         $portfolio = "portfolio";
+        $thumbnail = "thumbnail";
         $main_pic  = "main_pic";
         $thumbnail1 = "thumbnail1";
         $thumbnail2 = "thumbnail2";
@@ -149,14 +176,24 @@ public function delete_portfolio($id) {
             Storage::disk('local')->exists('/public/' . $portfolio . '/' . $thumbnail1 . '/' . $this->getUserDir() . '/' . $file->type . '/' . $file->floor_plan_code . '.' . $file->extension) &&
             Storage::disk('local')->exists('/public/' . $portfolio . '/' . $thumbnail2 . '/' . $this->getUserDir() . '/' . $file->type . '/' . $file->floor_plan_code . '.' . $file->extension) &&
             Storage::disk('local')->exists('/public/' . $portfolio . '/' . $thumbnail3 . '/' . $this->getUserDir() . '/' . $file->type . '/' . $file->floor_plan_code . '.' . $file->extension) &&
-            Storage::disk('local')->exists('/public/' . $portfolio . '/' . $thumbnail4 . '/' . $this->getUserDir() . '/' . $file->type . '/' . $file->floor_plan_code . '.' . $file->extension)
+            Storage::disk('local')->exists('/public/' . $portfolio . '/' . $thumbnail4 . '/' . $this->getUserDir() . '/' . $file->type . '/' . $file->floor_plan_code . '.' . $file->extension) && 
+            Storage::disk('local')->exists('/public/' . $thumbnail . '/' . $main_pic . '/' . $this->getUserDir() . $file->floor_plan_code . '.' . $file->extension) &&
+            Storage::disk('local')->exists('/public/' . $thumbnail . '/' . $thumbnail1 . '/' . $this->getUserDir() . $file->floor_plan_code . '.' . $file->extension) &&
+            Storage::disk('local')->exists('/public/' . $thumbnail . '/' . $thumbnail2 . '/' . $this->getUserDir() . $file->floor_plan_code . '.' . $file->extension) &&
+            Storage::disk('local')->exists('/public/' . $thumbnail . '/' . $thumbnail3 . '/' . $this->getUserDir() . $file->floor_plan_code . '.' . $file->extension) &&
+            Storage::disk('local')->exists('/public/' . $thumbnail . '/' . $thumbnail4 . '/' . $this->getUserDir() . $file->floor_plan_code . '.' . $file->extension)
              ) {
             if (
                 Storage::disk('local')->delete('/public/' . $portfolio . '/' . $main_pic . '/' . $this->getUserDir() . '/' . $file->type . '/' . $file->floor_plan_code . '.' . $file->extension) &&
                 Storage::disk('local')->delete('/public/' . $portfolio . '/' . $thumbnail1 . '/' . $this->getUserDir() . '/' . $file->type . '/' . $file->floor_plan_code . '.' . $file->extension) &&
                 Storage::disk('local')->delete('/public/' . $portfolio . '/' . $thumbnail2 . '/' . $this->getUserDir() . '/' . $file->type . '/' . $file->floor_plan_code . '.' . $file->extension) &&
                 Storage::disk('local')->delete('/public/' . $portfolio . '/' . $thumbnail3 . '/' . $this->getUserDir() . '/' . $file->type . '/' . $file->floor_plan_code . '.' . $file->extension) &&
-                Storage::disk('local')->delete('/public/' . $portfolio . '/' . $thumbnail4 . '/' . $this->getUserDir() . '/' . $file->type . '/' . $file->floor_plan_code . '.' . $file->extension)
+                Storage::disk('local')->delete('/public/' . $portfolio . '/' . $thumbnail4 . '/' . $this->getUserDir() . '/' . $file->type . '/' . $file->floor_plan_code . '.' . $file->extension) &&
+                Storage::disk('local')->delete('/public/' . $thumbnail . '/' . $main_pic . '/' . $this->getUserDir() . $file->floor_plan_code . '.' . $file->extension) &&
+                Storage::disk('local')->delete('/public/' . $thumbnail . '/' . $thumbnail1 . '/' . $this->getUserDir() . $file->floor_plan_code . '.' . $file->extension) &&
+                Storage::disk('local')->delete('/public/' . $thumbnail . '/' . $thumbnail2 . '/' . $this->getUserDir() . $file->floor_plan_code . '.' . $file->extension) &&
+                Storage::disk('local')->delete('/public/' . $thumbnail . '/' . $thumbnail3 . '/' . $this->getUserDir() . $file->floor_plan_code . '.' . $file->extension) &&
+                Storage::disk('local')->delete('/public/' . $thumbnail . '/' . $thumbnail4 . '/' . $this->getUserDir() . $file->floor_plan_code . '.' . $file->extension)
              ) {
                 return response()->json($file->delete());
             }
@@ -193,6 +230,89 @@ public function delete_portfolio($id) {
          $msg = "Portfolio updated sucessfully!";
 
          return response()->json($msg);
+
+    }
+
+     public function change_interior_email(Request $request) {
+
+        $model = new User();
+
+        $current_email = $request['c_email'];
+
+        $user = $model::where('id', Auth::id())->first();
+
+        if($user['email'] == $current_email) {
+
+            $verify = password_verify($request['password'], $user['password']);
+
+            $new_email = $request['n_email'];
+
+            if($verify) {
+
+                $update_email = $model::where('id', Auth::id())
+                        ->update(['email' => $new_email]);
+
+                $msg = "Email changed successfully!";   
+                     
+                return response()->json($msg);
+
+            } else {
+
+                $msg = "Incorrect password!";   
+                     
+                return response()->json($msg);
+            }
+
+        }else {
+             $msg = "Email doesn't recognized!";   
+                     
+             return response()->json($msg);
+        }
+
+
+    }
+
+
+    public function change_interior_password(Request $request) {
+
+        $model = new User();
+
+        $user = $model::where('id', Auth::id())->first();
+
+        $verify = password_verify($request['c_password'], $user['password']);
+
+
+        if($request['n_password'] == $request['cn_password']) {
+
+            if($verify) {
+
+                    $hash = Hash::make($request['n_password']);
+
+                    $update_password = $model::where('id', Auth::id())
+                            ->update(['password' => $hash]);
+
+                    $msg = "Password changed successfully!";   
+                         
+                    return response()->json($msg);
+
+             
+            } else {
+
+                 $msg = "Password doesnt recognized!";   
+                         
+                 return response()->json($msg);
+            }
+
+
+        }else {
+
+                $msg = "Incorrect confirm password!";   
+                     
+                return response()->json($msg);
+        }
+
+
+
 
     }
